Fix some compile issues in gcc 4.3.3

+++ muduo/base/LogStream.cc
@@ -61,7 +61,7 @@ size_t convertHex(char buf[], uintptr_t value)
 
   do
   {
-    int lsd = i % 16;
+    int lsd = static_cast<int>(i % 16);
     i /= 16;
     *p++ = digitsHex[lsd];
   } while (i != 0);
